missing include, stack overflow

+++ src/level/point.h
@@ -2,14 +2,17 @@
 #define ROGUECRAFT_POINT_H
 
 
+#include <stdint.h>
+
+
 #define point_new(y, x) (Point) {y, x}
 
 #define point_eq(p1, p2) (p1.x == p2.x && p1.y == p2.y)
 
 
 typedef struct {
-    int y;
-    int x;
+    int16_t y;
+    int16_t x;
 } Point;
