+++ src/player/player_movement.c
@@ -63,7 +63,7 @@ void player_move(Player *player, Input input)
     }
 }
 
-bool player_can_see(Player *player, uint32_t y, uint32_t x)
+bool player_can_see(Player *player, uint16_t y, uint16_t x)
 {
     Point pos = player->position.current;
     uint32_t dist = square(pos.x - x) + square(pos.y - y);
